requirements.txt
requests==2.26.0
beautifulsoup4==4.10.0
pandas==1.5.3
snowflake-connector-python
snowflake-connector-python[pandas]